back/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
dj-database-url==2.0.0
Django==3.2.2
django-cors-headers==3.4.0
djangorestframework==3.11.2
idna==3.4
Pillow==8.1.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.0
urllib3==2.0.3
